# packages/contracts/openapi/standards-knowledge-layer.openapi.yaml
openapi: 3.1.0
info:
  title: AIM RBI Standards & Regulatory Knowledge Layer
  version: 0.1.0
  description: >
    Contract surface for controlled standards metadata, private document ingestion,
    draft requirement extraction, human review, rule activation, and assistive
    engineering analysis. This contract never exposes licensed full text or final
    engineering authority outputs.
servers:
  - url: http://localhost:3001
    description: NestJS governance API
  - url: http://localhost:8000
    description: FastAPI engineering and extraction service
...
paths:
  /standards:
    get:
      tags: [Standards]
      summary: Search standards library metadata
      parameters:
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: publisher
          schema: { $ref: "#/components/schemas/StandardPublisher" }
        - in: query
          name: status
          schema: { $ref: "#/components/schemas/DocumentStatus" }
      responses:
        "200":
          description: Matching metadata records
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/StandardDocument" }
    post:
      tags: [Standards]
      summary: Register standards metadata
      description: Metadata-only registration. Binary standards remain in private storage.
      responses:
        "201": { description: Created }
  /standards/{standardId}:
    get:
      tags: [Standards]
      summary: Get one standard metadata record
      parameters:
        - $ref: "#/components/parameters/StandardId"
      responses:
        "200": { description: Standard metadata }
...
components:
  parameters:
    StandardId:
      in: path
      name: standardId
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
    StandardPublisher:
      type: string
      enum: [API, ISO, ASME, IEC, AMPP_NACE, ESDM, MIGAS, KEMNAKER, internal, client, other]
    DocumentStatus:
      type: string
      enum: [uploaded, parsing, parsed, extraction_pending, extracted, review_pending, approved, active, superseded, rejected, archived]
...
    StandardDocument:
      type: object
      required: [id, code, normalizedCode, title, publisher, status, metadataOnly]
      properties:
        id: { type: string }
        code: { type: string }
        normalizedCode: { type: string }
        title: { type: string }
        publisher: { $ref: "#/components/schemas/StandardPublisher" }
        status: { $ref: "#/components/schemas/DocumentStatus" }
        metadataOnly: { type: boolean }
        documentStorageUri:
          type: string
          description: Private URI/path only; never raw file content.